requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
python-dotenv>=1.0.0
firecrawl-py>=0.0.1
pydantic>=2.0.0
aiohttp>=3.9.0
